requirements.txt
azure-cognitiveservices-speech
openai
msal